pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "benchmark-vlm-cpu"
version = "0.1.0"
description = "Benchmarking suite for small Vision-Language Models running on CPU"
requires-python = ">=3.10"
dependencies = [
    # Core inference
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "transformers>=4.45.0",
    "accelerate>=0.33.0",
    "sentencepiece>=0.2.0",
    "pillow>=10.0.0",
    # Quantization / faster CPU inference
    "optimum>=1.21.0",
    # Metrics
    "nltk>=3.8",
    "rouge-score>=0.1.2",
    "evaluate>=0.4.0",
    # Utilities
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.12.0",
    "pandas>=2.0",
    "datasets>=2.20.0",
    "huggingface-hub>=0.24.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-benchmark", "ruff", "mypy"]
llama-cpp = ["llama-cpp-python>=0.2.0"]

[project.scripts]
vlm-bench = "benchmark.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["benchmark*"]

[tool.ruff]
line-length = 100
target-version = "py310"
